+++ src/hooks/use-checkbox-group/useCheckboxGroupProps.tsx
@@ -1,49 +1,64 @@
+import { UseFieldOptions } from "form-atoms";
 import { useAtomValue } from "jotai";
-import { ChangeEvent, useEffect, useState } from "react";
+import { ChangeEvent, useMemo, useRef } from "react";
 
 import {
-  UseMultiSelectFieldProps as UseCheckboxGroupFieldProps,
+  type UseMultiSelectFieldProps as UseCheckboxGroupFieldProps,
   useFieldProps,
 } from "../";
-import { ZodArrayField, ZodFieldValue } from "../../fields";
+import type { ZodArrayField, ZodFieldValue } from "../../fields";
 
-export const useCheckboxGroupFieldProps = <
-  Option,
-  Field extends ZodArrayField,
->({
-  field,
-  options,
-  getValue,
-}: UseCheckboxGroupFieldProps<Option, Field>) => {
+export const useCheckboxGroupFieldProps = <Option, Field extends ZodArrayField>(
+  { field, options, getValue }: UseCheckboxGroupFieldProps<Option, Field>,
+  fieldOptions?: UseFieldOptions<ZodFieldValue<Field>>,
+) => {
   const atom = useAtomValue(field);
   const fieldValue = useAtomValue(atom.value);
-  const [value, setValue] = useState<number[]>(() =>
-    fieldValue.map((activeOption) => options.indexOf(activeOption)),
+  const optionValues = useMemo(
+    () => options.map(getValue),
+    [getValue, options],
   );
+
+  const prevValue = useRef(fieldValue);
+
+  const activeIndexes = useRef<number[]>(
+    fieldValue.map((activeOption) => optionValues.indexOf(activeOption)),
+  );
+
+  if (prevValue.current != fieldValue) {
+    /**
+     * The field was set from outside via initialValue, reset action, or set manually.
+     * Recompute the indexes.
+     **/
+    activeIndexes.current = fieldValue.map((activeOption) =>
+      optionValues.indexOf(activeOption),
+    );
+  }
 
   const getEventValue = (event: ChangeEvent<HTMLInputElement>) => {
     const index = parseInt(event.currentTarget.value);
-    const values = event.currentTarget.checked
-      ? [...value, index]
-      : value.filter((val) => val != index);
+    const nextIndexes = event.currentTarget.checked
+      ? [...activeIndexes.current, index]
+      : activeIndexes.current.filter((val) => val != index);
 
-    setValue(values);
+    activeIndexes.current = nextIndexes;
 
-    const activeOptions = values
-      .map((idx) => options[idx])
-      .filter(Boolean) as Option[];
+    const nextValues = nextIndexes.map((index) => optionValues[index]);
 
-    return activeOptions.map(getValue) as ZodFieldValue<Field>;
+    /**
+     * When user change event happened, we set the value.
+     * On the next render when the fieldValue is updated, we can skip calculating the activeIndexes.
+     */
+    prevValue.current = nextValues;
+
+    return nextValues as ZodFieldValue<Field>;
   };
 
-  useEffect(() => {
-    if (fieldValue.length === 0 && value.length !== 0) {
-      // reset local state, when form was reset
-      setValue([]);
-    }
-  }, [fieldValue]);
+  const props = useFieldProps<Field, HTMLInputElement>(
+    field,
+    getEventValue,
+    fieldOptions,
+  );
 
-  const props = useFieldProps<Field, HTMLInputElement>(field, getEventValue);
-
-  return { ...props, value };
+  return { ...props, value: activeIndexes.current };
 };