+++ src/hooks/use-multiselect-field-props/useMultiSelectFieldProps.tsx
@@ -1,5 +1,6 @@
+import { UseFieldOptions } from "form-atoms";
 import { useAtomValue } from "jotai";
-import { ChangeEvent, useCallback, useMemo, useState } from "react";
+import { ChangeEvent, useCallback, useMemo, useRef } from "react";
 import { ArrayCardinality, ZodAny, ZodArray } from "zod";
 
 import { UseOptionsProps, useFieldProps } from "..";
@@ -18,32 +19,57 @@ export type ZodArrayFieldValue<Field> =
     ? Value["_output"]
     : never;
 
-export const useMultiSelectFieldProps = <Option, Field extends ZodArrayField>({
-  field,
-  options,
-  getValue,
-}: UseMultiSelectFieldProps<Option, Field>) => {
+export const useMultiSelectFieldProps = <Option, Field extends ZodArrayField>(
+  { field, options, getValue }: UseMultiSelectFieldProps<Option, Field>,
+  fieldOptions?: UseFieldOptions<ZodFieldValue<Field>>,
+) => {
   const atom = useAtomValue(field);
   const fieldValue = useAtomValue(atom.value);
-  // TODO: getValue should be useMemo dependency, currently we asume it's stable
-  const values = useMemo(() => options.map(getValue), [options]);
-  const [value, setValue] = useState<string[]>(() =>
-    fieldValue.map((value) => `${values.indexOf(value)}`),
+  const optionValues = useMemo(
+    () => options.map(getValue),
+    [getValue, options],
   );
 
+  const prevValue = useRef(fieldValue);
+
+  const activeIndexes = useRef<string[]>(
+    fieldValue.map((activeOption) => `${optionValues.indexOf(activeOption)}`),
+  );
+
+  if (prevValue.current != fieldValue) {
+    /**
+     * The field was set from outside via initialValue, reset action, or set manually.
+     * Recompute the indexes.
+     **/
+    activeIndexes.current = fieldValue.map(
+      (activeOption) => `${optionValues.indexOf(activeOption)}`,
+    );
+  }
   const getEventValue = useCallback((event: ChangeEvent<HTMLSelectElement>) => {
-    const value = [...event.currentTarget.options]
+    const nextIndexes = [...event.currentTarget.options]
       .filter((option) => option.selected)
       .map((option) => option.value);
 
-    setValue(value);
+    activeIndexes.current = nextIndexes;
 
-    const nextValue = value.map((idx) => values[parseInt(idx)]);
+    const nextValues = nextIndexes.map(
+      (index) => optionValues[parseInt(index)],
+    );
 
-    return nextValue as ZodFieldValue<Field>;
+    /**
+     * When user change event happened, we set the value.
+     * On the next render when the fieldValue is updated, we can skip calculating the activeIndexes.
+     */
+    prevValue.current = nextValues;
+
+    return nextValues as ZodFieldValue<Field>;
   }, []);
 
-  const props = useFieldProps<Field, HTMLSelectElement>(field, getEventValue);
+  const props = useFieldProps<Field, HTMLSelectElement>(
+    field,
+    getEventValue,
+    fieldOptions,
+  );
 
-  return { ...props, value };
+  return { ...props, value: activeIndexes.current };
 };