+++ src/hooks/use-number-field-props/useNumberFieldProps.ts
@@ -1,7 +1,8 @@
+import { UseFieldOptions } from "form-atoms";
 import { ChangeEvent } from "react";
 
 import { FieldProps, useFieldProps } from "../";
-import { type NumberField } from "../../fields/";
+import type { NumberField, NumberFieldValue } from "../../fields";
 
 export type NumberFieldProps = FieldProps<NumberField>;
 
@@ -11,9 +12,12 @@ const getNumber = (event: ChangeEvent<HTMLInputElement>) => {
   return Number.isNaN(valueAsNumber) ? undefined : valueAsNumber;
 };
 
-export const useNumberFieldProps = (field: NumberField) => {
+export const useNumberFieldProps = (
+  field: NumberField,
+  options?: UseFieldOptions<NumberFieldValue>,
+) => {
   // transform undefined to "" to make the number input empty
-  const { value = "", ...props } = useFieldProps(field, getNumber);
+  const { value = "", ...props } = useFieldProps(field, getNumber, options);
 
   return { ...props, value };
 };